Move files to ONNX folder
